+++ src/lib.rs
@@ -7,7 +7,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! tokio-retry = "0.4"
+//! tokio-retry2 = "0.4"
 //! ```
 //!
 //! # Example
@@ -15,7 +15,7 @@
 //! ```rust,no_run
 
 //! use tokio_retry2::Retry;
-//! use tokio_retry2::strategy::{ExponentialBackoff, jitter};
+//! use tokio_retry2::strategy::ExponentialBackoff;
 //!
 //! async fn action() -> Result<u64, ()> {
 //!     // do some real-world stuff here...
@@ -25,13 +25,35 @@
 //! # #[tokio::main]
 //! # async fn main() -> Result<(), ()> {
 //! let retry_strategy = ExponentialBackoff::from_millis(10)
-//!     .map(jitter) // add jitter to delays
 //!     .take(3);    // limit to 3 retries
 //!
 //! let result = Retry::spawn(retry_strategy, action).await?;
 //! # Ok(())
 //! # }
 //! ```
+//!
+//! # Features
+//! `[jitter]``
+//!
+//! To use jitter, add this to your Cargo.toml
+//!
+//! ```toml
+//! [dependencies]
+//! tokio-retry2 = { version = "0.4", features = ["jitter"] }
+//! ```
+//!
+//! # Example
+//!
+//! ```rust,no_run
+//! use tokio_retry2::Retry;
+//! use tokio_retry2::strategy::{ExponentialBackoff, jitter};
+//!
+//! let retry_strategy = ExponentialBackoff::from_millis(10)
+//!    .map(jitter) // add jitter to the retry interval
+//!    .take(3);    // limit to 3 retries
+//!
+//!
+//!
 
 #![allow(warnings)]