+++ src/lib.rs
@@ -15,7 +15,7 @@
 //! ```rust,no_run
 
 //! use tokio_retry2::Retry;
-//! use tokio_retry2::strategy::ExponentialBackoff;
+//! use tokio_retry2::strategy::{ExponentialBackoff, MaxInterval};
 //!
 //! async fn action() -> Result<u64, ()> {
 //!     // do some real-world stuff here...
@@ -25,6 +25,7 @@
 //! # #[tokio::main]
 //! # async fn main() -> Result<(), ()> {
 //! let retry_strategy = ExponentialBackoff::from_millis(10)
+//!     .max_interval(10000) // set max interval to 10 seconds
 //!     .take(3);    // limit to 3 retries
 //!
 //! let result = Retry::spawn(retry_strategy, action).await?;
@@ -46,14 +47,18 @@
 //!
 //! ```rust,no_run
 //! use tokio_retry2::Retry;
-//! use tokio_retry2::strategy::{ExponentialBackoff, jitter};
+//! use tokio_retry2::strategy::{ExponentialBackoff, jitter, MaxInterval};
 //!
 //! let retry_strategy = ExponentialBackoff::from_millis(10)
 //!    .map(jitter) // add jitter to the retry interval
+//!    .max_interval(10000) // set max interval to 10 seconds
 //!    .take(3);    // limit to 3 retries
+//!````
 //!
-//!
-//!
+//! ### NOTE:
+//! The time spent executing an action does not affect the intervals between
+//! retries. Therefore, for long-running functions it's a good idea to set up a deadline,
+//! to place an upper bound on the strategy execution time.
 
 #![allow(warnings)]