+++ src/strategy/mod.rs
@@ -3,10 +3,12 @@ mod fibonacci_backoff;
 mod fixed_interval;
 #[cfg(feature = "jitter")]
 mod jitter;
+mod max_interval;
 
 pub use self::exponential_backoff::ExponentialBackoff;
 pub use self::fibonacci_backoff::FibonacciBackoff;
 pub use self::fixed_interval::FixedInterval;
+pub use self::max_interval::MaxInterval;
 
 #[cfg(feature = "jitter")]
 pub use self::jitter::jitter;