+++ src/strategy/jitter.rs
@@ -1,4 +1,4 @@
-use std::time::Duration;
+use tokio::time::Duration;
 
 pub fn jitter(duration: Duration) -> Duration {
     duration.mul_f64(rand::random::<f64>())